+++ sys/win32/win_syscon.cpp
@@ -100,7 +100,7 @@ static LONG WINAPI ConWndProc( HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lPara
 		case WM_CLOSE:
 			if ( cvarSystem->IsInitialized() && com_skipRenderer.GetBool() ) {
 				cmdString = Mem_CopyString( "quit" );
-				Sys_QueEvent( 0, SE_CONSOLE, 0, 0, strlen( cmdString ) + 1, cmdString );
+                Sys_QueEvent(0, SE_CONSOLE, 0, 0, static_cast<int>(strlen(cmdString)) + 1, cmdString);
 			} else if ( s_wcd.quitOnClose ) {
 				PostQuitMessage( 0 );
 			} else {
@@ -138,7 +138,7 @@ static LONG WINAPI ConWndProc( HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lPara
 					PostQuitMessage( 0 );
 				} else {
 					cmdString = Mem_CopyString( "quit" );
-					Sys_QueEvent( 0, SE_CONSOLE, 0, 0, strlen( cmdString ) + 1, cmdString );
+                    Sys_QueEvent(0, SE_CONSOLE, 0, 0, static_cast<int>(strlen(cmdString)) + 1, cmdString);
 				}
 			} else if ( wParam == CLEAR_ID ) {
 				SendMessage( s_wcd.hwndBuffer, EM_SETSEL, 0, -1 );
